+++ apps/website/src/_state/component-statuses.ts
@@ -27,7 +27,7 @@ export const statusByComponent: ComponentKitsStatuses = {
     RadioGroup: ComponentStatus.Draft,
     Separator: ComponentStatus.Beta,
     Skeleton: ComponentStatus.Beta,
-    Sonner: ComponentStatus.Draft,
+    Toast: ComponentStatus.Draft,
     Tabs: ComponentStatus.Beta,
     Textarea: ComponentStatus.Draft,
   },
