+++ apps/website/src/routes/docs/styled/menu.md
@@ -28,6 +28,6 @@
 - [RadioGroup](/docs/styled/radio-group)
 - [Separator](/docs/styled/separator)
 - [Skeleton](/docs/styled/skeleton)
-- [Sonner](/docs/styled/sonner)
+- [Toast](/docs/styled/toast)
 - [Tabs](/docs/styled/tabs)
 - [Textarea](/docs/styled/textarea)
