+++ packages/kit-styled/src/index.ts
@@ -13,7 +13,7 @@ export * from './components/popover/popover';
 export * from './components/radio-group/radio-group';
 export * from './components/separator/separator';
 export * from './components/skeleton/skeleton';
-export * from './components/sonner/sonner';
+export * from './components/toast/toast';
 export * from './components/tabs/tabs';
 export * from './components/textarea/textarea';
 export * from './components/toggle/toggle';
